organizaed the routes of react-router each on with it file as individual component and imported in AppRouter.js
