+++ src/components/HelpPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const HelpPage = () => (
+  <div>
+    This is from my help component
+  </div>
+);
+
+export default HelpPage;
